pyproject.toml
[project]
name = "cross-persona-introspection"
version = "0.1.0"
description = "Framework for testing persona-dependent introspection in LLMs"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.36",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "tqdm>=4.60",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ipykernel"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
